+++ backend/data/src/main/kotlin/io/tolgee/development/testDataBuilder/TestDataService.kt
@@ -86,9 +86,7 @@ class TestDataService(
     entityManager.clear()
 
     executeInNewTransaction(transactionManager) {
-      saveAllMtCreditBuckets(builder)
       saveProjectData(builder)
-
       finalize()
     }
 
@@ -120,6 +118,7 @@ class TestDataService(
   private fun saveOrganizationDependants(builder: TestDataBuilder) {
     saveOrganizationRoles(builder)
     saveOrganizationAvatars(builder)
+    saveAllMtCreditBuckets(builder)
   }
 
   private fun saveOrganizationAvatars(builder: TestDataBuilder) {
@@ -192,7 +191,7 @@ class TestDataService(
 
   private fun saveTranslationDependants(translationBuilders: List<TranslationBuilder>) {
     val translationComments = translationBuilders.flatMap { it.data.comments.map { it.self } }
-    translationCommentService.createAll(translationComments)
+    translationCommentService.saveAll(translationComments)
   }
 
   private fun saveTranslations(builder: ProjectBuilder): List<TranslationBuilder> {