+++ e2e/cypress/e2e/import/importApplication.cy.ts
@@ -33,12 +33,8 @@ describe('Import application', () => {
         'French',
         'English',
       ]);
-      cy.gcy('translations-table-cell')
-        .contains('What a text')
-        .should('be.visible');
-      cy.gcy('translations-table-cell')
-        .contains('What a french text')
-        .should('be.visible');
+      cy.contains('What a text').should('exist');
+      cy.contains('What a french text').should('exist');
     }
   );
 });